refactor: 머지충돌

+++ src/main/java/in/koreatech/koin/domain/shop/repository/ShopRepository.java
@@ -30,5 +30,5 @@ public interface ShopRepository extends Repository<Shop, Integer> {
 
     List<Shop> findAll();
 
-    void deleteByOwnerId(Long ownerId);
+    void deleteByOwnerId(Integer ownerId);
 }
